+++ Dolar blue oficial/src/com/lucianosimo/dolar/model/OldDolarValues.java
@@ -5,30 +5,30 @@ import com.google.gson.annotations.SerializedName;
 public class OldDolarValues {
 	
 	@SerializedName("old_oficial_compra")
-	private double old_oficial_compra;
+	private float old_oficial_compra;
 	
 	@SerializedName("old_oficial_venta")
-	private double old_oficial_venta;
+	private float old_oficial_venta;
 	
 	@SerializedName("old_blue_compra")
-	private double old_blue_compra;
+	private float old_blue_compra;
 	
 	@SerializedName("old_blue_venta")
-	private double old_blue_venta;
+	private float old_blue_venta;
 	
-	public double getOldOficialCompra() {
+	public float getOldOficialCompra() {
 		return old_oficial_compra;
 	}
 	
-	public double getOldOficialVenta() {
+	public float getOldOficialVenta() {
 		return old_oficial_venta;
 	}
 	
-	public double getOldBlueCompra() {
+	public float getOldBlueCompra() {
 		return old_blue_compra;
 	}
 	
-	public double getOldBlueVenta() {
+	public float getOldBlueVenta() {
 		return old_blue_venta;
 	}
